+++ SplunkAgent/Sources/SplunkAgent/Public API/API-1.0-EndpointConfiguration.swift
@@ -76,7 +76,7 @@ public struct EndpointConfiguration: Codable, Equatable {
     private static func realmUrl(for realm: String, path: String) -> URL? {
         var urlComponents = URLComponents()
         urlComponents.scheme = "https"
-        urlComponents.host = "rum-ingest.\(realm).signalfx.com"
+        urlComponents.host = "rum-ingest.\(realm).observability.splunkcloud.com"
         urlComponents.path = path
 
         return urlComponents.url