+++ SplunkNavigation/Sources/SplunkNavigation/Model/NavigationModel.swift
@@ -22,6 +22,8 @@ actor NavigationModel {
 
     private(set) var moduleEnabled: Bool = true
     private(set) var screenName: String = "unknown"
+    private(set) var isManualScreenName = false
+
     private(set) var navigations: [ObjectIdentifier: NavigationPair] = [:]
     private(set) var agentVersion: String?
 
@@ -37,6 +39,10 @@ actor NavigationModel {
 
     func update(screenName: String) {
         self.screenName = screenName
+    }
+
+    func update(isManualScreenName: Bool) {
+        self.isManualScreenName = isManualScreenName
     }