+++ CHANGELOG.md
@@ -9,13 +9,13 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 
+* Added public references to sessionWillResetNotification and sessionDidResetNotification
 * Added optional support to capture network headers.
 * Session Replay now captures text content in wireframes for UIKit applications.
 
 ### Changed
 
 * Updated endpoint from signalfx.com to observability.splunkcloud.com
-
 
 ## [2.2.1] - 2026-03-19