+++ CHANGELOG.md
@@ -7,6 +7,8 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [2.2.2] - 2026-04-15
+
 ### Added
 
 * Added public references to sessionWillResetNotification and sessionDidResetNotification